feat: add update cards module
